Edit and Rename Dump of Database (mtg_tracker.sql)
The original sql file, 'group9_db.sql', was created using mysqldump
WITHOUT the `--databases` flag and, thus, did not contain `CREATE` and
`USE` commands for the `group9` database, which are useful when
importing the sql file on a new server that does not currently contain
any version of the database (to see the exact lines added, use `diff`).

The file was also renamed from 'group9_db.sql' to 'mtg_tracker.sql' to
highlight the fact that the file was changed and to prevent confusion,
as well as to better match the naming scheme of the project.